removed double rendering of Nav bar
